feat(impl): [#518] split up test step def classes and update todos
