+++ irs-cucumber-tests/src/test/java/org/eclipse/tractusx/irs/cucumber/E2ETestStepDefinitions.java
@@ -30,23 +30,18 @@ import static org.awaitility.Awaitility.await;
 import java.io.File;
 import java.io.IOException;
 import java.time.Duration;
-import java.time.OffsetDateTime;
-import java.util.ArrayList;
 import java.util.HashSet;
-import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
 import java.util.UUID;
 import java.util.concurrent.TimeUnit;
-import java.util.function.Predicate;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import io.cucumber.java.After;
-import io.cucumber.java.Before;
 import io.cucumber.java.DataTableType;
 import io.cucumber.java.PendingException;
 import io.cucumber.java.Scenario;
@@ -55,8 +50,6 @@ import io.cucumber.java.en.Given;
 import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
 import io.restassured.http.ContentType;
-import io.restassured.specification.RequestSpecification;
-import org.apache.commons.lang3.StringUtils;
 import org.eclipse.tractusx.irs.component.BatchOrderCreated;
 import org.eclipse.tractusx.irs.component.BatchOrderResponse;
 import org.eclipse.tractusx.irs.component.BatchResponse;
@@ -72,16 +65,10 @@ import org.eclipse.tractusx.irs.component.enums.BatchStrategy;
 import org.eclipse.tractusx.irs.component.enums.BomLifecycle;
 import org.eclipse.tractusx.irs.component.enums.Direction;
 import org.eclipse.tractusx.irs.component.enums.JobState;
-import org.eclipse.tractusx.irs.policystore.models.CreatePoliciesResponse;
-import org.eclipse.tractusx.irs.policystore.models.CreatePolicyRequest;
-import org.eclipse.tractusx.irs.policystore.models.UpdatePolicyRequest;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 
 public class E2ETestStepDefinitions {
-
-    public static final String URL_IRS_POLICIES = "/irs/policies";
-    public static final String QUERYPARAM_BUSINESS_PARTNER_NUMBERS = "businessPartnerNumbers";
 
     private RegisterJob.RegisterJobBuilder registerJobBuilder;
 
@@ -96,15 +83,12 @@ public class E2ETestStepDefinitions {
     private BatchOrderResponse batchOrderResponse;
     private BatchResponse batchResponse;
 
-    private ObjectMapper objectMapper;
+    private final ObjectMapper objectMapper;
 
     private AuthenticationProperties authProperties;
-    private AuthenticationProperties.AuthenticationPropertiesBuilder authenticationPropertiesBuilder;
+    private final AuthenticationProperties.AuthenticationPropertiesBuilder authenticationPropertiesBuilder;
 
-    private Map<String, ArrayList<LinkedHashMap<String, ?>>> bpnToPoliciesMap;
-
-    @Before
-    public void setup() {
+    public E2ETestStepDefinitions() {
         registerJobBuilder = RegisterJob.builder();
         registerBatchOrderBuilder = RegisterBatchOrder.builder();
         authenticationPropertiesBuilder = AuthenticationProperties.builder();
@@ -521,171 +505,6 @@ public class E2ETestStepDefinitions {
     @After("@INTEGRATION_TEST")
     public void addJobIdToResult(Scenario scenario) {
         scenario.attach(jobId.toString(), MediaType.TEXT_PLAIN_VALUE, "jobId");
-    }
-
-    @Given("Clean up all test policies with prefix {string}")
-    public void cleanupPoliciesWithPrefix(final String policyIdPrefix) {
-        fetchPolicyIdsByPrefix(policyIdPrefix).forEach(this::cleanupPolicy);
-    }
-
-    @When("I fetch all policies")
-    public void iFetchAllPolicies() {
-        this.bpnToPoliciesMap = fetchAllPolicies();
-    }
-
-    @When("I fetch policies for BPN {string}")
-    public void iFetchPoliciesForBpn(final String bpn) {
-        this.bpnToPoliciesMap = fetchPoliciesForBpn(bpn);
-    }
-
-    @Then("the BPN {string} should have the following policies:")
-    public void theBpnShouldHaveTheFollowingPolicies(final String bpn, final List<String> policyIds) {
-        final List<String> policyIdsForBpn = PolicyTestHelper.extractPolicyIdsForBpn(bpnToPoliciesMap, bpn).toList();
-        assertThat(policyIdsForBpn).containsAll(policyIds);
-    }
-
-    @When("I delete the following policies:")
-    public void iDeleteTheFollowingPolicies(final List<String> policyIds) {
-        for (final String policyId : policyIds) {
-            cleanupPolicy(policyId);
-        }
-    }
-
-    @Then("the BPN {string} should have {int} policies having policyId starting with {string}")
-    public void theBpnShouldHaveTheFollowingPolicies(final String bpn, final int numPolicies, final String prefix) {
-        final List<String> policyIdsForBpn = PolicyTestHelper.extractPolicyIdsForBpn(bpnToPoliciesMap, bpn)
-                                                             .filter(startingWith(prefix))
-                                                             .toList();
-        assertThat(policyIdsForBpn).hasSize(numPolicies);
-    }
-
-    private static Predicate<String> startingWith(final String prefix) {
-        return s -> StringUtils.startsWith(s, prefix);
-    }
-
-    @When("I perform update policy {string}, BPN {string}, validUntil {string}")
-    public void iPerformUpdate(final String policyId, final String bpn, final String validUntil) {
-        updatePolicies(List.of(policyId), List.of(bpn), validUntil);
-    }
-
-    @When("I add policyId {string} to given BPNs using validUntil {string}:")
-    public void iAddPolicyIdToGivenBpns(final String policyId, final String validUntil,
-            List<String> businessPartnerNumbers) {
-        updatePolicies(List.of(policyId), businessPartnerNumbers, validUntil);
-    }
-
-    @Then("the BPN {string} should have a policy with policyId {string} and validUntil {string}")
-    @SuppressWarnings({ "rawtypes",
-                        "unchecked"
-    })
-    public void theBpnShouldHaveTheExpectedPolicyWithValidUntil(final String bpn, final String policyId,
-            final String validUntil) {
-        final List<LinkedHashMap> policies = PolicyTestHelper.extractPoliciesForBpn(bpnToPoliciesMap, bpn).toList();
-        final List<LinkedHashMap> policiesFiltered = policies.stream()
-                                                             .filter(p -> p.get("policyId").equals(policyId))
-                                                             .toList();
-        assertThat(policiesFiltered).hasSize(1);
-        assertThat(policiesFiltered.get(0)).containsEntry("policyId", policyId) //
-                                           .containsEntry("validUntil", validUntil);
-    }
-
-    @Given("a policy with policyId {string} for BPN {string} and validUntil {string}")
-    public void iRegisterAPolicy(final String policyId, final String bpn, final String validUntil) {
-        final String policyJson = PolicyTestHelper.policyTemplate.formatted(policyId);
-        final CreatePoliciesResponse response = registerPolicyForBpn(policyJson, bpn, validUntil);
-        assertThat(response.policyId()).isEqualTo(policyId);
-    }
-
-    private void cleanupPolicy(final String policyId) {
-
-        final int status = givenAuthentication().pathParam("policyId", policyId)
-                                                .when()
-                                                .delete(URL_IRS_POLICIES + "/{policyId}")
-                                                .then()
-                                                .log()
-                                                .all()
-                                                .extract()
-                                                .statusCode();
-
-        assertThat(List.of(HttpStatus.OK.value(), HttpStatus.NOT_FOUND.value())).describedAs(
-                "Should either return status 200 OK or 404 NOT_FOUND").contains(status);
-    }
-
-    private List<String> fetchPolicyIdsByPrefix(final String policyIdPrefix) {
-        final Map<String, ArrayList<LinkedHashMap<String, ?>>> bpnToPoliciesMap = fetchAllPolicies();
-        return PolicyTestHelper.extractPolicyIdsStartingWith(bpnToPoliciesMap, policyIdPrefix).toList();
-    }
-
-    @SuppressWarnings("unchecked")
-    private Map<String, ArrayList<LinkedHashMap<String, ?>>> fetchAllPolicies() {
-
-        return givenAuthentication().when()
-                                    .get(URL_IRS_POLICIES)
-                                    .then()
-                                    .log()
-                                    .all()
-                                    .statusCode(HttpStatus.OK.value())
-                                    .extract()
-                                    .body()
-                                    .as(Map.class);
-    }
-
-    @SuppressWarnings("unchecked")
-    private Map<String, ArrayList<LinkedHashMap<String, ?>>> fetchPoliciesForBpn(final String bpn) {
-
-        return givenAuthentication().queryParam(QUERYPARAM_BUSINESS_PARTNER_NUMBERS, bpn)
-                                    .when()
-                                    .get(URL_IRS_POLICIES)
-                                    .then()
-                                    .log()
-                                    .all()
-                                    .statusCode(HttpStatus.OK.value())
-                                    .extract()
-                                    .body()
-                                    .as(Map.class);
-    }
-
-    private CreatePoliciesResponse registerPolicyForBpn(final String policyJson, final String bpn,
-            final String validUntil) {
-
-        final var createPolicyRequest = CreatePolicyRequest.builder()
-                                                           .validUntil(OffsetDateTime.parse(validUntil))
-                                                           .businessPartnerNumber(bpn)
-                                                           .payload(PolicyTestHelper.jsonFromString(policyJson))
-                                                           .build();
-        return givenAuthentication().contentType(ContentType.JSON)
-                                    .body(createPolicyRequest)
-                                    .when()
-                                    .post(URL_IRS_POLICIES)
-                                    .then()
-                                    .log()
-                                    .all()
-                                    .statusCode(HttpStatus.CREATED.value())
-                                    .extract()
-                                    .as(CreatePoliciesResponse.class);
-    }
-
-    private void updatePolicies(final List<String> policyIds, final List<String> businessPartnerNumbers,
-            final String validUntil) {
-
-        final var updatePolicyRequest = UpdatePolicyRequest.builder()
-                                                           .policyIds(policyIds)
-                                                           .businessPartnerNumbers(businessPartnerNumbers)
-                                                           .validUntil(OffsetDateTime.parse(validUntil))
-                                                           .build();
-        givenAuthentication().contentType(ContentType.JSON)
-                             .body(updatePolicyRequest)
-                             .when()
-                             .put(URL_IRS_POLICIES)
-                             .then()
-                             .log()
-                             .all()
-                             .statusCode(HttpStatus.OK.value());
-    }
-
-    private RequestSpecification givenAuthentication() {
-        authProperties = authenticationPropertiesBuilder.build();
-        return given().log().all().spec(authProperties.getNewAuthenticationRequestSpecification());
     }
 
 }
